% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit.R
\name{extract_fit}
\alias{extract_fit}
\title{Extract fit statistics of lavaan objects}
\usage{
extract_fit(lavaan_object, details = FALSE)
}
\arguments{
\item{lavaan_object}{Vector, lavaan object(s)}

\item{details}{Logical, if TRUE return all fit statistics.
By default this is set to FALSE, a selection (chisq, npar, aic, bic, cfi, rmsea, srmr) of fit statistics is returned.}
}
\value{
This function returns a tibble.
}
\description{
Extract fit statistics of lavaan objects
}
