% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcsm_data}
\alias{lcsm_data}
\title{Longitudinal dataset with repeated measures of two constructs}
\format{
A longitudinal dataset in wide format:
\itemize{
\item \code{id}: ID variable, unique identifier for each person
\item \code{x1}: x value at time point 1
\item \code{x2}: x value at time point 2
\item \code{x3}: x value at time point 3
\item \code{x4}: x value at time point 4
\item \code{x5}: x value at time point 5
\item \code{y1}: y value at time point 1
\item \code{y2}: y value at time point 2
\item \code{y3}: y value at time point 3
\item \code{y4}: y value at time point 4
\item \code{y5}: y value at time point 5
}
}
\usage{
data(lcsm_data)
}
\description{
Example dataset with 5 repeated measures of two constructs to illustrate how the package works.
}
\examples{
# Load data into global environment
data(lcsm_data)
}
\keyword{dataset}
