% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lcsm.R
\name{plot_lcsm}
\alias{plot_lcsm}
\title{Plot latent change score model path diagram}
\usage{
plot_lcsm(object, layout, what = "col", whatLabels = "est",
  edge.width = 1, node.width = 1, border.width = 1, fixedStyle = 1,
  freeStyle = 1, residuals = FALSE, label.scale = FALSE,
  sizeMan = 5, sizeLat = 5, intercepts = FALSE, fade = FALSE,
  nCharNodes = 0, nCharEdges = 0)
}
\arguments{
\item{object}{}

\item{layout}{}

\item{what}{See \code{semPlot}}

\item{whatLabels}{}

\item{edge.width}{}

\item{node.width}{}

\item{border.width}{}

\item{fixedStyle}{}

\item{freeStyle}{}

\item{residuals}{}

\item{label.scale}{}

\item{sizeMan}{}

\item{sizeLat}{}

\item{intercepts}{}

\item{fade}{}

\item{nCharNodes}{}

\item{nCharEdges}{}
}
\value{
Plot
}
\description{
Plot latent change score model path diagram
}
\examples{
TODO
}
