% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot individual trajectories}
\usage{
plot_trajectories(data, id_var, var_list, line_colour = "blue",
  point_colour = "black", line_alpha = 0.2, point_alpha = 0.2,
  smooth = FALSE, smooth_method = "loess", smooth_se = FALSE,
  xlab = "X", ylab = "Y", scale_x_num = FALSE,
  scale_x_num_start = 1, random_sample_frac = 1, title_n = FALSE,
  connect_missing = TRUE)
}
\arguments{
\item{data}{Dataset in wide format.}

\item{id_var}{String, specifying id variable.}

\item{var_list}{Vector, specifying variable names to be plotted in sequential order.}

\item{line_colour}{String, specifying colour of lines.}

\item{point_colour}{String, specifying, colour of points.}

\item{line_alpha}{Numeric, specifying alpha of lines.}

\item{point_alpha}{Numeric, specifying alpha of points.}

\item{smooth}{Logical, add line of average using method = \code{smooth_method}.}

\item{smooth_method}{String, specifying method to be used for calculating average line, see  \code{ggplot2}.}

\item{smooth_se}{Locical, specifying whether to add standard error of average line or not.}

\item{xlab}{String for x axis label.}

\item{ylab}{String for y axis label.}

\item{scale_x_num}{Logical, if \code{TRUE} print sequential numbers starting from 1 as x axis labels, if \code{FALSE} use variable names.}

\item{scale_x_num_start}{Numeric, if \code{scale_x_num == TRUE} this is the starting value of the x axis.}

\item{random_sample_frac}{The fraction of rows to select (from wide dataset), default is set to 1 (100 percent) of the sample.}

\item{title_n}{Logical, speciying whether to print title with number and percentage of cases used for the plot.}

\item{connect_missing}{Logical, speciying whether to connect points by \code{id_var} across missing values.}
}
\value{
ggplot2 object
}
\description{
Plot individual trajectories
}
