% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_vars.R
\name{rename_vars}
\alias{rename_vars}
\title{Rename variables for univariate and bivariate models}
\usage{
rename_vars(data, x_vars, y_vars)
}
\arguments{
\item{data}{Dataframe in wide format}

\item{x_vars}{List of variables measuring first construct}

\item{y_vars}{List of variables measuring second construct}
}
\value{
Dataframe in wide format with renamed variables
}
\description{
Rename variables for univariate and bivariate models
}
