% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lcsm_data.R
\name{sim_bi_lcsm_data}
\alias{sim_bi_lcsm_data}
\title{Simulate data from a univariate latent change score model}
\usage{
sim_bi_lcsm_data(timepoints, model_x, model_x_param = NULL, model_y,
  model_y_param = NULL, coupling, coupling_param = NULL,
  sample.nobs = 500, na_x_pct = 0, na_y_pct = 0, ..., var_x = "x",
  var_y = "y", change_letter_x = "g", change_letter_y = "j",
  return_lavaan_syntax = FALSE, return_lavaan_syntax_string = FALSE)
}
\arguments{
\item{timepoints}{See \link[lcstools]{specify_lavaan_bi_model}}

\item{model_x}{See \link[lcstools]{specify_lavaan_bi_model}}

\item{model_x_param}{List, specifying parameter estimates for the LCS model that has been specified in the argument '\code{model_x}':
\itemize{
\item{\strong{\code{gamma_lx1}}}: Mean of latent true scores x (Intercept),
\item{\strong{\code{sigma2_lx1}}}: Variance of latent true scores x,
\item{\strong{\code{sigma2_ux}}}: Variance of observed scores x,
\item{\strong{\code{alpha_g2}}}: Mean of change factor (g2),
\item{\strong{\code{alpha_g3}}}: Mean of change factor (g3),
\item{\strong{\code{sigma2_g2}}}: Variance of change factor (g2).
\item{\strong{\code{sigma2_g3}}}: Variance of change factor (g3),
\item{\strong{\code{sigma_g2lx1}}}: Covariance of change factor (g2) with the initial true score x (lx1),
\item{\strong{\code{sigma_g3lx1}}}: Covariance of change factor (g3) with the initial true score x (lx1),
\item{\strong{\code{sigma_g2g3}}}: Covariance of change factors (g2 and g2),
\item{\strong{\code{phi_x}}}: Autoregression of change scores x.
}}

\item{model_y}{See \link[lcstools]{specify_lavaan_bi_model}}

\item{model_y_param}{List, specifying parameter estimates for the LCS model that has been specified in the argument '\code{model_y}':
\itemize{
\item{\strong{\code{gamma_ly1}}}: Mean of latent true scores y (Intercept),
\item{\strong{\code{sigma2_ly1}}}: Variance of latent true scores y,
\item{\strong{\code{sigma2_uy}}}: Variance of observed scores y,
\item{\strong{\code{alpha_j2}}}: Mean of change factor (j2),
\item{\strong{\code{alpha_j3}}}: Mean of change factor (j3),
\item{\strong{\code{sigma2_j2}}}: Variance of change factor (j2).
\item{\strong{\code{sigma2_j3}}}: Variance of change factor (j3),
\item{\strong{\code{sigma_j2ly1}}}: Covariance of change factor (j2) with the initial true score x (ly1),
\item{\strong{\code{sigma_j3ly1}}}: Covariance of change factor (j3) with the initial true score x (ly1),
\item{\strong{\code{sigma_j2j3}}}: Covariance of change factors (j2 and j2),
\item{\strong{\code{phi_y}}}: Autoregression of change scores y.
}}

\item{coupling}{See \link[lcstools]{specify_lavaan_bi_model}}

\item{coupling_param}{List, specifying parameter estimates coupling parameters that have been specified in the argument '\code{coupling}':
\itemize{
\item{\strong{\code{sigma_su}}}: Covariance of residuals x and y,
\item{\strong{\code{sigma_ly1lx1}}}: Covariance of intercepts x and y,
\item{\strong{\code{sigma_g2ly1}}}: Covariance of change factor x (g2) with the initial true score y (ly1),
\item{\strong{\code{sigma_g3ly1}}}: Covariance of change factor x (g3) with the initial true score y (ly1),
\item{\strong{\code{sigma_j2lx1}}}: Covariance of change factor y (j2) with the initial true score x (lx1),
\item{\strong{\code{sigma_j3lx1}}}: Covariance of change factor y (j3) with the initial true score x (lx1),
\item{\strong{\code{sigma_j2g2}}}: Covariance of change factors y (j2) and x (g2),
\item{\strong{\code{sigma_j2g3}}}: Covariance of change factors y (j2) and x (g3),
\item{\strong{\code{sigma_j3g2}}}: Covariance of change factors y (j3) and x (g2),,
\item{\strong{\code{delta_con_xy}}}: Change score x (t) determined by true score y (t),
\item{\strong{\code{delta_con_yx}}}: Change score y (t) determined by true score x (t),
\item{\strong{\code{delta_lag_xy}}}: Change score x (t) determined by true score y (t-1),
\item{\strong{\code{delta_lag_yx}}}: Change score y (t) determined by true score x (t-1),
\item{\strong{\code{xi_con_xy}}}: Change score x (t) determined by change score y (t),
\item{\strong{\code{xi_con_yx}}}: Change score y (t) determined by change score x (t),
\item{\strong{\code{xi_lag_xy}}}: Change score x (t) determined by change score y (t-1),
\item{\strong{\code{xi_lag_yx}}}: Change score y (t) determined by change score x (t-1)
}}

\item{sample.nobs}{Numeric, number of cases to be simulated, see \link[lcstools]{specify_lavaan_uni_model}}

\item{na_x_pct}{Numeric, percentage of random missing values in the simulated dataset [0,1]}

\item{na_y_pct}{Numeric, percentage of random missing values in the simulated dataset [0,1]}

\item{...}{Arguments to be passed on to \link[lavaan]{simulateData}}

\item{var_x}{See \link[lcstools]{specify_lavaan_bi_model}}

\item{var_y}{See \link[lcstools]{specify_lavaan_bi_model}}

\item{change_letter_x}{See \link[lcstools]{specify_lavaan_bi_model}}

\item{change_letter_y}{See \link[lcstools]{specify_lavaan_bi_model}}

\item{return_lavaan_syntax}{Logical, if TRUE return the lavaan syntax used for simulating data, looking beautiful using \link[base]{cat}}

\item{return_lavaan_syntax_string}{Logical, if return_lavaan_syntax == TRUE and return_lavaan_syntax_string == TRUE return the lavaan syntax as one ugly string}
}
\description{
Simulate data from a univariate latent change score model
}
