% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lcsm_data.R
\name{sim_uni_lcsm_data}
\alias{sim_uni_lcsm_data}
\title{Simulate data from a univariate latent change score model}
\usage{
sim_uni_lcsm_data(timepoints, model, var = "x", change_letter = "j",
  sample.nobs = 500, ..., return_lavaan_syntax = FALSE,
  return_lavaan_syntax_string = FALSE, lavaan_uni_model = NULL,
  uni_model_param = NULL)
}
\arguments{
\item{timepoints}{See \link[lcstools]{specify_lavaan_uni_model}}

\item{model}{See \link[lcstools]{specify_lavaan_uni_model}}

\item{var}{See \link[lcstools]{specify_lavaan_uni_model}}

\item{change_letter}{See \link[lcstools]{specify_lavaan_uni_model}}

\item{...}{Arguments to be passed on to \link[lavaan]{simulateData}}

\item{lavaan_uni_model}{String, lavaan model syntax with "labals" for parameters}

\item{uni_model_param}{Placeholder, this might become an option where to scpecify parameter estimates}
}
\description{
Simulate data from a univariate latent change score model
}
