% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_lavaan_bi_model.R
\name{specify_lavaan_bi_model}
\alias{specify_lavaan_bi_model}
\title{Specify lavaan model for bivariate latent change score models}
\usage{
specify_lavaan_bi_model(timepoints, var_x, model_x, var_y, model_y,
  coupling, change_letter_x = "g", change_letter_y = "j")
}
\arguments{
\item{timepoints}{Number of timepoints.}

\item{var_x}{Vector, specifying variables measuring one construct of the model.}

\item{model_x}{List, specifying model specifications (logical) for variables specified in \code{var_x}.
\itemize{
\item{\code{alpha_constant}}{ (Constant change factor)},
\item{\code{alpha_piecewise}}{ (Piecewise constant change factors)},
\item{\code{alpha_piecewise_num}}{ (Changepoint of piecewise constant change factors)},
\item{\code{alpha_linear}}{ (Linear change factor)},
\item{\code{beta}}{ (Proportional change factor)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}

\item{var_y}{Vector, specifying variables measuring another construct of the model.}

\item{model_y}{List, specifying model specifications (logical) for variables specified in \code{var_y}.
\itemize{
\item{\code{alpha_constant}}{ (Constant change factor)},
\item{\code{alpha_piecewise}}{ (Piecewise constant change factors)},
\item{\code{alpha_piecewise_num}}{ (Changepoint of piecewise constant change factors)},
\item{\code{alpha_linear}}{ (Linear change factor)},
\item{\code{beta}}{ (Proportional change factor)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}

\item{coupling}{List, specifying coupling parameters.
\itemize{
\item{\code{coupling_piecewise}}{ (Piecewise coupling parameters)},
\item{\code{coupling_piecewise_num}}{ (Changepoint of piecewise coupling parameters)},
\item{\code{delta_xy}}{ (True score y predicting subsequent change score x)},
\item{\code{delta_yx}}{ (True score x predicting subsequent change score y)},
\item{\code{xi_xy}}{ (Change score y predicting subsequent change score x)},
\item{\code{xi_yx}}{ (Change score x predicting subsequent change score y)}.
}}

\item{change_letter_x}{String, specifying letter to be used as change factor for construct x in lavaan syntax.}

\item{change_letter_y}{String, specifying letter to be used as change factor for construct y in lavaan syntax.}
}
\value{
Lavaan model syntax including comments.
}
\description{
Specify lavaan model for bivariate latent change score models
}
