% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_uni_lcsm}
\alias{data_uni_lcsm}
\title{Longitudinal dataset with repeated measures of one constructs}
\format{
A longitudinal dataset in wide format:
\itemize{
  \item{id}{: ID variable, unique identifier for each person}
  \item{x1}{: x value at time point 1}
  \item{x2}{: x value at time point 2}
  \item{x3}{: x value at time point 3}
  \item{x4}{: x value at time point 4}
  \item{x5}{: x value at time point 5}
  \item{x6}{: x value at time point 6}
  \item{x7}{: x value at time point 7}
  \item{x8}{: x value at time point 8}
  \item{x9}{: x value at time point 9}
  \item{x10}{: x value at time point 10}
}
}
\usage{
data(data_uni_lcsm)
}
\description{
Example dataset with repeated measures of one constructs to illustrate how the package works.
}
\examples{
# Load data into global environment
data(data_uni_lcsm)
}
\keyword{dataset}
