% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit.R
\name{extract_fit}
\alias{extract_fit}
\title{Extract fit statistics of lavaan objects}
\usage{
extract_fit(lavaan_object, details = FALSE)
}
\arguments{
\item{lavaan_object}{Lavaan object(s). To extract fit from multiple lavaan objects use c().}

\item{details}{Show detailed fit information.}
}
\value{
This function returns a tibble.
}
\description{
Extract fit statistics of lavaan objects
}
