% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param.R
\name{extract_param}
\alias{extract_param}
\title{Extract labeled parameters of lavaan objects}
\usage{
extract_param(lavaan_object, printp = FALSE)
}
\arguments{
\item{lavaan_object}{Lavaan object.}

\item{printp}{If TRUE convert into easily readable p-values.}
}
\value{
This function returns a tibble with labeled parameters.
}
\description{
Extract labeled parameters of lavaan objects
}
