% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bi_lcsm.R
\name{fit_bi_lcsm}
\alias{fit_bi_lcsm}
\title{Fit bivariate latent change score models}
\usage{
fit_bi_lcsm(data, var_x, var_y, model_x, model_y, coupling,
  mimic = "Mplus", estimator = "MLR", missing = "FIML",
  export_model_syntax = FALSE)
}
\arguments{
\item{data}{Wide dataset.}

\item{var_x}{List of variables measuring one construct of the model.}

\item{var_y}{List of variables measuring another construct of the model.}

\item{model_x}{List of model specifications for variables specified in \code{var_x}.
\itemize{
\item{\code{alpha}}{ (Constant change)},
\item{\code{beta}}{ (Proportional change)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}

\item{model_y}{List of model specifications for variables specified in \code{var_y}.
\itemize{
\item{\code{alpha}}{ (Constant change)},
\item{\code{beta}}{ (Proportional change)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}

\item{coupling}{List of model specifications for coupling parameters.
\itemize{
\item{\code{delta_xy}}{ (True score y predicting subsequent change score x)},
\item{\code{delta_yx}}{ (True score x predicting subsequent change score y)},
\item{\code{xi_xy}}{ (Change score y predicting subsequent change score x)},
\item{\code{xi_yx}}{ (Change score x predicting subsequent change score y)}.
}}
}
\value{
This function returns a lavaan class object.
}
\description{
Fit bivariate latent change score models.
}
