% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcsm.R
\name{fit_bi_lcsm}
\alias{fit_bi_lcsm}
\title{Fit bivariate latent change score models}
\usage{
fit_bi_lcsm(data, var_x, var_y, model_x, model_y, coupling,
  mimic = "Mplus", estimator = "MLR", missing = "FIML",
  return_lavaan_syntax = FALSE, return_lavaan_syntax_string = FALSE,
  ...)
}
\arguments{
\item{data}{Wide dataset.}

\item{var_x}{List of variables measuring one construct of the model.}

\item{var_y}{List of variables measuring another construct of the model.}

\item{model_x}{List of model specifications (logical) for variables specified in \code{var_x}.
\itemize{
\item{\code{alpha_constant}}{ (Constant change factor)},
\item{\code{alpha_piecewise}}{ (Piecewise constant change factors)},
\item{\code{alpha_piecewise_num}}{ (Changepoint of piecewise constant change factors)},
\item{\code{alpha_linear}}{ (Linear change factor)},
\item{\code{beta}}{ (Proportional change factor)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}

\item{model_y}{List of model specifications for variables specified in \code{var_y}.
\itemize{
\item{\code{alpha_constant}}{ (Constant change factor)},
\item{\code{alpha_piecewise}}{ (Piecewise constant change factors)},
\item{\code{alpha_piecewise_num}}{ (Changepoint of piecewise constant change factors)},
\item{\code{alpha_linear}}{ (Linear change factor)},
\item{\code{beta}}{ (Proportional change factor)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}

\item{coupling}{List of model specifications (logical) for coupling parameters.
\itemize{
\item{\code{coupling_piecewise}}{ (Piecewise coupling parameters)},
\item{\code{coupling_piecewise_num}}{ (Changepoint of piecewise coupling parameters)},
\item{\code{delta_xy}}{ (True score y predicting subsequent change score x)},
\item{\code{delta_yx}}{ (True score x predicting subsequent change score y)},
\item{\code{xi_xy}}{ (Change score y predicting subsequent change score x)},
\item{\code{xi_yx}}{ (Change score x predicting subsequent change score y)}.
}}

\item{mimic}{See \link[lavaan]{lavaan}.}

\item{estimator}{See \link[lavaan]{lavaan}.}

\item{missing}{See \link[lavaan]{lavaan}.}

\item{return_lavaan_syntax}{Logical, if TRUE return the lavaan syntax used for simulating data, looking beautiful using \link[base]{cat}}

\item{return_lavaan_syntax_string}{Logical, if return_lavaan_syntax == TRUE and return_lavaan_syntax_string == TRUE return the lavaan syntax as one ugly string}

\item{...}{Additional arguments to be passed to \link[lavaan]{lavaan}.}
}
\value{
This function returns a lavaan class object.
}
\description{
Fit bivariate latent change score models.
}
\examples{
# Fit 
fit_bi_lcsm(data = data_bi_lcsm, 
            var_x = names(data_bi_lcsm)[2:11], 
            var_y = names(data_bi_lcsm)[12:21],
            model_x = list(alpha_constant = TRUE, beta = TRUE, phi = FALSE),
            model_y = list(alpha_constant = TRUE, beta = TRUE, phi = TRUE),
            coupling = list(delta_lag_xy = TRUE, xi_lag_yx = TRUE))
}
