% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bi_lcsm.R
\name{fit_bi_lcsm}
\alias{fit_bi_lcsm}
\title{Fit bivariate latend change score models}
\usage{
fit_bi_lcsm(data, y, x, model, mimic = "Mplus", estimator = "MLR",
  missing = "FIML")
}
\arguments{
\item{data}{Wide dataset.}

\item{y}{List of 6 to 10 variables measuring the other construct of the model.}

\item{x}{List of 6 to 10 variables measuring one construct of the model.}

\item{model}{Model specifications as described in Grimm2012 and Grimm2017. The model names M1a to M4b referr to the abbreviations used in Grimm2012 to describe these models. The options are:
\itemize{
\item{\code{bi_lcsm_no_coupling}}{ (M1a)},
\item{\code{bi_lcsm_x_on_deltay}}{ (M2a)},
\item{\code{bi_lcsm_y_on_deltax}}{ (M3a)},
\item{\code{bi_lcsm_dual_change}}{ (M4a)},
\item{\code{bi_lcsm_dual_change_delta_within}}{ (M1b)},
\item{\code{bi_lcsm_dual_change_delta_xy}}{ (M2b)},
\item{\code{bi_lcsm_dual_change_delta_yx}}{ (M3b)},
\item{\code{bi_lcsm_dual_change_delta_full}}{ (M4b)}.
}}
}
\value{
This function returns a lavaan class object.
}
\description{
Fit bivariate latend change score models
}
