% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_uni_lcsm.R
\name{fit_uni_lcsm}
\alias{fit_uni_lcsm}
\title{Fit univariate latent change score models}
\usage{
fit_uni_lcsm(data, var, model, mimic = "Mplus", estimator = "MLR",
  missing = "FIML")
}
\arguments{
\item{data}{Wide dataset.}

\item{var}{List of variables measuring construct of the model.}

\item{model}{List of model specifications for variables specified in \code{var}.
\itemize{
\item{\code{alpha}}{ (Constant change)},
\item{\code{beta}}{ (Proportional change)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}
}
\value{
This function returns a lavaan class object.
}
\description{
Fit univariate latent change score models.
}
