% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_uni_lcsm.R
\name{fit_uni_lcsm}
\alias{fit_uni_lcsm}
\title{Fit univariate latent change score models}
\usage{
fit_uni_lcsm(data, var, model, mimic = "Mplus", estimator = "MLR",
  missing = "FIML", export_model_syntax = FALSE,
  name_model_syntax = "lavaan_model_syntax", ...)
}
\arguments{
\item{data}{Wide dataset.}

\item{var}{List of variables measuring construct of the model.}

\item{model}{List of model specifications (logical) for variables specified in \code{var}.
\itemize{
\item{\code{alpha_constant}}{ (Constant change factor)},
\item{\code{alpha_piecewise}}{ (Piecewise constant change factors)},
\item{\code{alpha_piecewise_num}}{ (Changepoint of piecewise constant change factors},
\item{\code{alpha_linear}}{ (Linear change factor)},
\item{\code{beta}}{ (Proportional change factor)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}

\item{export_model_syntax}{Export lavaan model syntax of specified model to global environment as object named 'lavaan_model_syntax'. Name of this object can be specified using \code{name_model_syntax}.}

\item{name_model_syntax}{String, if \code{export_model_syntax} = TRUE, name of object containing lavaan model syntax.}
}
\value{
This function returns a lavaan class object.
}
\description{
Fit univariate latent change score models.
}
