% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcsm_data}
\alias{lcsm_data}
\title{Longitudinal dataset with repeated measures of two constructs}
\format{
A longitudinal dataset in wide format:
\itemize{
  \item{id}{: ID variable, unique identifier for each person}
  \item{x1}{: x value at time point 1}
  \item{x2}{: x value at time point 2}
  \item{x3}{: x value at time point 3}
  \item{x4}{: x value at time point 4}
  \item{x5}{: x value at time point 5}
  \item{y1}{: y value at time point 1}
  \item{y2}{: y value at time point 2}
  \item{y3}{: y value at time point 3}
  \item{y4}{: y value at time point 4}
  \item{y5}{: y value at time point 5}
}
}
\usage{
data(lcsm_data)
}
\description{
Example dataset with 5 repeated measures of two constructs to illustrate how the package works.
}
\examples{
# Load data into global environment
data(lcsm_data)
}
\keyword{dataset}
