% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lcsm.R
\name{plot_lcsm}
\alias{plot_lcsm}
\title{Plot latent change score model path diagram}
\usage{
plot_lcsm(lavaan_object, layout = NULL, lcsm = c("univariate",
  "bivariate"), curve_covar = 0.5, what = "col", whatLabels = "est",
  edge.width = 1, node.width = 1, border.width = 1, fixedStyle = 1,
  freeStyle = 1, residuals = FALSE, label.scale = FALSE,
  sizeMan = 3, sizeLat = 5, intercepts = FALSE, fade = FALSE,
  nCharNodes = 0, nCharEdges = 0, edge.label.cex = 0.5,
  DoNotPlot = TRUE, ...)
}
\arguments{
\item{lavaan_object}{lavaan object.}

\item{layout}{Matrix specifying location of manifest and latent variables.}

\item{curve_covar}{see \code{semPlot}.}

\item{what}{See \code{semPlot}.}

\item{whatLabels}{see \code{semPlot}.}

\item{edge.width}{see \code{semPlot}.}

\item{node.width}{see \code{semPlot}.}

\item{border.width}{see \code{semPlot}.}

\item{fixedStyle}{see \code{semPlot}.}

\item{freeStyle}{see \code{semPlot}.}

\item{residuals}{see \code{semPlot}.}

\item{label.scale}{see \code{semPlot}.}

\item{sizeMan}{see \code{semPlot}.}

\item{sizeLat}{see \code{semPlot}.}

\item{intercepts}{see \code{semPlot}.}

\item{fade}{see \code{semPlot}.}

\item{nCharNodes}{see \code{semPlot}.}

\item{nCharEdges}{see \code{semPlot}.}

\item{edge.label.cex}{see \code{semPlot}.}

\item{DoNotPlot}{see \code{semPlot}}

\item{...}{Other functions passed on to \code{semPlot}.}
}
\value{
Plot
}
\description{
Plot latent change score model path diagram
}
