% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot individual trajectories}
\usage{
plot_trajectories(data, id_var, var_list, line_colour = "blue",
  point_colour = "black", line_alpha = 0.2, point_alpha = 0.2,
  smooth = FALSE, smooth_method = "loess", smooth_se = FALSE,
  xlab = "X", ylab = "Y")
}
\arguments{
\item{data}{Dataset in wide format}

\item{id_var}{String of id variable}

\item{var_list}{List of variables to be plotted}

\item{line_colour}{Colour of lines}

\item{point_colour}{Colour of points}

\item{line_alpha}{Alpha of lines}

\item{point_alpha}{Alpha of points}

\item{smooth}{Logical, add line of average}

\item{smooth_method}{Method for calculating average}

\item{smooth_se}{Locical, add standard error of average line}

\item{xlab}{String for x-Axis label}

\item{ylab}{String for y-Axis label}
}
\value{
ggplot2 object
}
\description{
Plot individual trajectories
}
\examples{
TODO
}
