% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cases.R
\name{select_bi_cases}
\alias{select_bi_cases}
\title{Select cases based on minimum number of available session scores on two longitudinal measures}
\usage{
select_bi_cases(data, id_var, var_list_x, var_list_y, min_count_x,
  min_count_y)
}
\arguments{
\item{data}{TODO}

\item{id_var}{TODO}

\item{var_list_x}{TODO}

\item{var_list_y}{TODO}

\item{min_count_x}{TODO}

\item{min_count_y}{TODO}
}
\value{
tibble
}
\description{
TODO: Describe function
}
