% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cases.R
\name{select_uni_cases}
\alias{select_uni_cases}
\title{Select cases based on minimum number of available session scores on one longitudinal measure}
\usage{
select_uni_cases(data, id_var, var_list, min_count,
  return_id_only = FALSE)
}
\arguments{
\item{data}{TODO}

\item{id_var}{TODO}

\item{var_list}{TODO}

\item{min_count}{TODO}

\item{return_id_only}{TODO}
}
\value{
tibble
}
\description{
Select cases based on minimum number of available session scores on one longitudinal measure
}
\examples{

}
