% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_lavaan_bi_model.R
\name{specify_lavaan_bi_model}
\alias{specify_lavaan_bi_model}
\title{Specify lavaan model for bivariate latent change score models}
\usage{
specify_lavaan_bi_model(timepoints, var_x, model_x, var_y, model_y,
  coupling, change_letter_x = "g", change_letter_y = "j")
}
\arguments{
\item{timepoints}{Number if timepoints.}

\item{model_x}{List of model specifications for variables specified in \code{var_x}.
\itemize{
\item{\code{alpha}}{ (Constant change)},
\item{\code{beta}}{ (Proportional change)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}

\item{model_y}{List of model specifications for variables specified in \code{var_y}.
\itemize{
\item{\code{alpha}}{ (Constant change)},
\item{\code{beta}}{ (Proportional change)},
\item{\code{phi}}{ (Autoregression of change scores)}.
}}

\item{coupling}{List of model specifications for coupling parameters.
\itemize{
\item{\code{delta_xy}}{ (True score y predicting subsequent change score x)},
\item{\code{delta_yx}}{ (True score x predicting subsequent change score y)},
\item{\code{xi_xy}}{ (Change score y predicting subsequent change score x)},
\item{\code{xi_yx}}{ (Change score x predicting subsequent change score y)}.
}}
}
\value{
Lavaan model syntax.
}
\description{
Specify lavaan model for bivariate latent change score models
}
\examples{
TODO.


}
