% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_lcsm_lavaan_syntax.R
\name{specify_lavaan_uni_model}
\alias{specify_lavaan_uni_model}
\title{Specify lavaan model for univariate latent change score models}
\usage{
specify_lavaan_uni_model(timepoints, model, var, change_letter)
}
\arguments{
\item{timepoints}{Number if timepoints.}

\item{model}{List of model specifications (logical) for the variables specified in \code{variable}.
\itemize{
\item{\code{alpha_constant}: Constant change factor},
\item{\code{alpha_piecewise}: Piecewise constant change factors},
\item{\code{alpha_piecewise_num}: Changepoint of piecewise constant change factors},
\item{\code{alpha_linear}: Linear change factor},
\item{\code{beta}: Proportional change factor},
\item{\code{phi}: Autoregression of change scores}.
}}

\item{var}{String, specifying letter to be used for of variables (Usually x or y).}

\item{change_letter}{String, specifying letter to be used for change factor (Usually g or j).}
}
\value{
Lavaan model syntax including comments.
}
\description{
TODO: Describe function
}
