% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_lavaan_uni_model.R
\name{specify_lavaan_uni_model}
\alias{specify_lavaan_uni_model}
\title{Specify lavaan model for univariate latent change score models}
\usage{
specify_lavaan_uni_model(timepoints, model, variable, change_letter)
}
\arguments{
\item{timepoints}{Number if timepoints.}

\item{model}{List of parameters to be estimated in the univariate lcsm.
\itemize{
\item{\code{alpha}}{ (Constant change)},
\item{\code{beta}}{ (Proportional change)},
\item{\code{phi}}{ (Autoregression of change scores)},
\item{\code{alpha_piecewise}}{ (Piecewise constant change)},
\item{\code{alpha_piecewise_num}}{ (Change point of piecewise constant change)}.
}}

\item{variable}{String with letter of variable (Usually x or y).}

\item{change_letter}{String with letter (Usually g or j).}
}
\value{
Lavaan model syntax.
}
\description{
Specify lavaan model for univariate latent change score models
}
\examples{
TODO. 

}
